refactor(group): group node render move to group-plugin
