+++ apps/demo-free-layout/src/components/index.ts
@@ -2,4 +2,3 @@ export * from './base-node';
 export * from './line-add-button';
 export * from './node-panel';
 export * from './comment';
-export * from './group';
