+++ apps/demo-free-layout/src/plugins/free-group-plugin/index.ts
@@ -1,2 +1,3 @@
 export { createFreeGroupPlugin } from './create-free-group-plugin';
 export { WorkflowGroupService } from './workflow-group-service';
+export { WorkflowGroupCommand } from './constant';
