+++ apps/demo-free-layout/src/plugins/free-group-plugin/render/components/background.tsx
