+++ apps/demo-free-layout/src/plugins/free-group-plugin/render/components/index.ts
@@ -1,2 +1,2 @@
 export { GroupNodeRender } from './node-render';
-export { GroupTools } from './tools';
+export { IconGroup } from './icon-group';
