+++ apps/demo-free-layout/src/plugins/free-group-plugin/render/index.ts
@@ -1,3 +1,4 @@
 import './index.css';
 
 export { GroupNodeRender } from './components';
+export { IconGroup } from './components';
